pyproject.toml
[tool.poetry]
name = "backend-we-eat"
version = "0.1.0"
description = ""
authors = ["Samuel Stegmeyer"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
motor = "^3.0.0"
httpx = "^0.23.0"
pydantic = "^1.10.2"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 180
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''